+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/qhybrid/NotificationHRConfiguration.java
@@ -1,18 +1,33 @@
 package nodomain.freeyourgadget.gadgetbridge.devices.qhybrid;
 
-import android.util.Log;
-
 import java.io.Serializable;
-
-import nodomain.freeyourgadget.gadgetbridge.service.devices.qhybrid.requests.misfit.PlayNotificationRequest;
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.util.zip.CRC32;
 
 public class NotificationHRConfiguration implements Serializable {
     private String packageName;
     private long id = -1;
+    private byte[] packageCrc;
 
     public NotificationHRConfiguration(String packageName, long id) {
         this.packageName = packageName;
         this.id = id;
+
+        CRC32 crc = new CRC32();
+        crc.update(packageName.getBytes());
+
+        this.packageCrc = ByteBuffer
+                .allocate(4)
+                .order(ByteOrder.LITTLE_ENDIAN)
+                .putInt((int) crc.getValue())
+                .array();
+    }
+
+    public NotificationHRConfiguration(String packageName, byte[] packageCrc, long id) {
+        this.id = id;
+        this.packageCrc = packageCrc;
+        this.packageName = packageName;
     }
 
     public String getPackageName() {
@@ -21,5 +36,9 @@ public class NotificationHRConfiguration implements Serializable {
 
     public long getId() {
         return id;
+    }
+
+    public byte[] getPackageCrc() {
+        return packageCrc;
     }
 }