+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/qhybrid/requests/fossil_hr/notification/NotificationFilterPutHRRequest.java
@@ -3,7 +3,6 @@ package nodomain.freeyourgadget.gadgetbridge.service.devices.qhybrid.requests.fo
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.util.ArrayList;
-import java.util.zip.CRC32;
 
 import nodomain.freeyourgadget.gadgetbridge.devices.qhybrid.NotificationHRConfiguration;
 import nodomain.freeyourgadget.gadgetbridge.service.devices.qhybrid.adapter.fossil.FossilWatchAdapter;
@@ -21,20 +20,16 @@ public class NotificationFilterPutHRRequest extends FilePutRequest {
     }
 
     private static byte[] createFile(NotificationHRConfiguration[] configs) {
-        ByteBuffer buffer = ByteBuffer.allocate(configs.length * 28);
+        int payloadLength = configs.length * 28;
+        ByteBuffer buffer = ByteBuffer.allocate(payloadLength);
         buffer.order(ByteOrder.LITTLE_ENDIAN);
 
         for (NotificationHRConfiguration config : configs) {
-            buffer.putShort((short) 28); //packet length
+            payloadLength = 26;
 
-            CRC32 crc = new CRC32();
-            crc.update(config.getPackageName().getBytes());
+            buffer.putShort((short) payloadLength); //packet length
 
-            byte[] crcBytes = ByteBuffer
-                    .allocate(4)
-                    .order(ByteOrder.LITTLE_ENDIAN)
-                    .putInt((int) crc.getValue())
-                    .array();
+            byte[] crcBytes = config.getPackageCrc();
 
             // 6 bytes
             buffer.put(PacketID.PACKAGE_NAME_CRC.id)
@@ -44,7 +39,7 @@ public class NotificationFilterPutHRRequest extends FilePutRequest {
             // 3 bytes
             buffer.put(PacketID.GROUP_ID.id)
                     .put((byte) 1)
-                    .put((byte) 2);
+                    .put((byte) 0);
 
             // 3 bytes
             buffer.put(PacketID.PRIORITY.id)